project3/requirements.txt
bentoml>=1.0.0
torch
torchvision
pillow
numpy
requests
